requirements.txt
albumentations==1.3.1
keras==2.15.0
matplotlib==3.7.1
numpy==1.25.2
pandas==2.0.3
Pillow==9.4.0
scikit-learn==1.2.2
tensorflow==2.15.0
segmentation-models==1.0.1
pydantic==2.7.1
fastapi==0.111.0
typing==3.7.4.3
uvicorn==0.29.0
streamlit==1.34.0